abort controller imitation
