+++ src/pages/SchedulerDayPC/components/MainContent.jsx
@@ -1,8 +1,9 @@
 import React from "react";
+
+import { HourMarkersColumn } from "../../../components/HourMarkersColumn";
+import { CurrentMomentRedLine } from "../../../components/CurrentMomentRedLine";
+import { HourLines } from "../../../components/HourLines";
 import { DayColumn } from "../../../components/DayColumn";
-import { CurrentMomentRedLine } from "../../../components/DayColumnList/components/CurrentMomentRedLine";
-import { HourLines } from "../../../components/DayColumnList/components/HourLines";
-import { HourMarkersColumn } from "../../../components/DayColumnList/components/HourMarkersColumn";
 import { DayPreview } from "./DayPreviewPlaceholder";
 
 export function MainContent({ dayDate, eventsInSelectedDay }) {
